+++ lib/src/navigation_pane_theme.dart
@@ -76,14 +76,14 @@ class PaneThemeData with Diagnosticable {
   /// [NavigationDestination] labels.
   ///
   /// You can use this to specify a different style when the label is selected.
-  final MaterialStateProperty<TextStyle?>? labelTextStyle;
+  final WidgetStateProperty<TextStyle?>? labelTextStyle;
 
   /// The theme to merge with the default icon theme for
   /// [NavigationDestination] icons.
   ///
   /// You can use this to specify a different icon theme when the icon is
   /// selected.
-  final MaterialStateProperty<IconThemeData?>? iconTheme;
+  final WidgetStateProperty<IconThemeData?>? iconTheme;
 
   /// Overrides the default value of [Pane.compactWidth].
   final double? compactWidth;
@@ -105,8 +105,8 @@ class PaneThemeData with Diagnosticable {
     Color? indicatorColor,
     ShapeBorder? indicatorShape,
     Size? indicatorSize,
-    MaterialStateProperty<TextStyle?>? labelTextStyle,
-    MaterialStateProperty<IconThemeData?>? iconTheme,
+    WidgetStateProperty<TextStyle?>? labelTextStyle,
+    WidgetStateProperty<IconThemeData?>? iconTheme,
     double? compactWidth,
     double? openWidth,
   }) {
@@ -154,13 +154,13 @@ class PaneThemeData with Diagnosticable {
       indicatorColor: Color.lerp(a?.indicatorColor, b?.indicatorColor, t),
       indicatorShape: ShapeBorder.lerp(a?.indicatorShape, b?.indicatorShape, t),
       indicatorSize: Size.lerp(a?.indicatorSize, a?.indicatorSize, t),
-      labelTextStyle: MaterialStateProperty.lerp<TextStyle?>(
+      labelTextStyle: WidgetStateProperty.lerp<TextStyle?>(
         a?.labelTextStyle,
         b?.labelTextStyle,
         t,
         TextStyle.lerp,
       ),
-      iconTheme: MaterialStateProperty.lerp<IconThemeData?>(
+      iconTheme: WidgetStateProperty.lerp<IconThemeData?>(
         a?.iconTheme,
         b?.iconTheme,
         t,
@@ -270,14 +270,14 @@ class PaneThemeData with Diagnosticable {
       ),
     );
     properties.add(
-      DiagnosticsProperty<MaterialStateProperty<TextStyle?>>(
+      DiagnosticsProperty<WidgetStateProperty<TextStyle?>>(
         'labelTextStyle',
         labelTextStyle,
         defaultValue: null,
       ),
     );
     properties.add(
-      DiagnosticsProperty<MaterialStateProperty<IconThemeData?>>(
+      DiagnosticsProperty<WidgetStateProperty<IconThemeData?>>(
         'iconTheme',
         iconTheme,
         defaultValue: null,