+++ CHANGELOG.md
@@ -1,3 +1,13 @@
+## [1.0.3]
+
+### Feat
+- Added `footers` property to Navigation Pane for displaying additional widgets after the main navigation items.
+- Added `additionalLending` feature to Navigation App Bar.
+
+### Fixed
+- Fixed a bug in the `PaneController`.
+- Migrates deprecated `MaterialState` and `MaterialStateProperty` to `WidgetState` and `WidgetStateProperty`.
+
 ## [1.0.1]
 
 ### Fixed